ooDialog - Invoke leaving() for the PropertySheetDialog
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/main/trunk@7744 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ extensions/platform/windows/oodialog/ooDialog.hpp
@@ -849,7 +849,7 @@ typedef struct _psdCSelf {
     char                *caption;
     uint32_t             pageCount;
     uint32_t             propSheetFlags;
-    int                  getResultValue;   // Storage for the return from PSM_GETRESULT
+    intptr_t             getResultValue;   // Storage for the return from PSM_GETRESULT
     bool                 modeless;
     bool                 isNotWizard;
     bool                 isWiz97;
